==> notebooks/12_load_features_hopsworks.ipynb <==
from dotenv import load_dotenv
import os
load_dotenv() 

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from datetime import datetime
import pandas as pd
from src.data_utils import load_and_process_taxi_data

from_year = 2024
# to_year = datetime.now().year
to_year = 2025
print(f"Download raw data from {from_year} to {to_year}")

rides = pd.DataFrame()
chunks = []
for year in range(from_year, to_year+1):

    rides_one_year = load_and_process_taxi_data(year)

    chunks.append(rides_one_year)
    break

# Concatenate all chunks at the end
rides = pd.concat(chunks, ignore_index=True)
print("Data loading complete.")

rides

rides.shape

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

ts_data.shape

ts_data.info()

ts_data

import hopsworks
import pandas as pd
from hsfs.feature_group import FeatureGroup

project = hopsworks.login()
feature_store = project.get_feature_store()
# Create the feature group
feature_group = feature_store.create_feature_group(
    name="time_series_6h_feature_group",
    version=1,
    description="6 hourly ride data",
    primary_key=[ "pickup_locatin_id", "pickup_hour"],  # Adjust primary key according to your dataset
    event_time="pickup_hour",  # Adjust event time according to your dataset
    online_enabled=True
)
print("Feature group created successfully!")

# feature_group = feature_store.create_feature_group(
#     name="time_series_hourly_feature_group",
#     version=1,
#     description="Feature group of hourly time series data",  # Use the new 'id' column as the primary key
#     online_enabled=True   # Enable online serving if needed
# )
# print("Feature group created with supported primary key for online serving.")

ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'])
ts_data['pickup_location_id'] = ts_data['pickup_location_id'].astype(int)
ts_data['rides'] = ts_data['rides'].astype(int)

FEATURE_GROUP_NAME = "time_series_6h_feature_group"
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time series data at hourly frequency",
    primary_key=['pickup_location_id', 'pickup_hour'],
    event_time=['pickup_hour']
)

ts_data.rename(columns={
    "pickup_hour": 'pickup_hour',
    "pickup_location_id": 'pickup_location_id',
    "rides": 'rides'
}, inplace=True)

ts_data_selected = ts_data[['pickup_hour', 'pickup_location_id', 'rides']]

feature_group.insert(ts_data_selected, write_options={"wait_for_job": False})

ts_data

import hopsworks 
project = hopsworks.login()

feature_store = project.get_feature_store(name='thanoojl_featurestore')

feature_group = fg = feature_store.get_feature_group('time_series_6h_feature_group', version=1)

feature_group.insert(ts_data, write_options={"wait_for_job": False})

df_memory_mb = rides.memory_usage(deep=True).sum() / (1024 * 1024)  
print(f"DataFrame size: {df_memory_mb:.2f} MB")

ts_data.info()
